backend/pyproject.toml
[project]
name = "itzuli-nlp"
version = "0.1.0"
description = "Basque NLP tools with Itzuli translation and Stanza morphological analysis"
readme = "../README.md"
requires-python = ">=3.10"
dependencies = [
    "dependencies>=7.7.1",
    "fastapi>=0.129.0",
    "fastmcp>=2.14.5",
    "ipython>=8.18.1",
    "itzuli>=1.1.0",
    "mcp>=1.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "stanza>=1.11.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-anyio>=0.0.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/itzuli_nlp"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--ignore-glob=*_manual.py"
